@@ -11,27 +11,26 @@
     "url": "https://github.com/openchemistry/mongochemclient"
   },
   "dependencies": {
-    "axios": "^0.17.0",
+    "axios": "^0.18.0",
     "d3": "^3.5.6",
     "enumify": "^1.0.4",
     "filesize": "^3.5.11",
     "font-awesome": "^4.7.0",
-    "jsonpath": "^0.2.12",
+    "jsonpath": "^1.0.0",
     "lodash": "^4.17.4",
-    "material-ui": "^0.18.3",
+    "material-ui": "^0.20.1",
     "molecule-3d-for-react": "git+https://github.com/cjh1/molecule-3d-for-react.git",
     "moment": "^2.19.1",
     "normalizr": "^3.2.2",
-    "postinstall-build": "^3.0.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
     "react-iframe": "^1.1.0",
     "react-redirect": "^1.0.0",
     "react-redux": "^5.0.4",
     "react-router": "^4.1.0",
     "react-router-redux": "^5.0.0-alpha.6",
-    "react-tap-event-plugin": "^2.0.1",
-    "redux": "^3.6.0",
+    "react-scripts": "1.1.4",
+    "redux": "^4.0.0",
     "redux-actions": "^2.0.2",
     "redux-form": "^7.1.2",
     "redux-form-material-ui": "^4.3.0",
@@ -41,77 +40,51 @@
     "universal-cookie": "^2.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "6.7.2",
-    "babel-core": "6.22.1",
-    "babel-eslint": "7.1.1",
-    "babel-jest": "18.0.0",
-    "babel-loader": "6.2.10",
+    "autoprefixer": "8.6.1",
+    "babel-core": "6.26.3",
+    "babel-eslint": "8.2.3",
+    "babel-jest": "23.0.1",
+    "babel-loader": "7.1.4",
     "babel-plugin-lodash": "^3.2.11",
     "babel-preset-es2016": "^6.3.13",
-    "babel-preset-react-app": "^2.2.0",
+    "babel-preset-react-app": "^3.1.1",
     "babel-runtime": "^6.20.0",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "connect-history-api-fallback": "1.3.0",
-    "cross-spawn": "4.0.2",
-    "css-loader": "0.26.1",
-    "detect-port": "1.1.0",
-    "dotenv": "2.0.0",
-    "eslint": "3.16.1",
-    "eslint-config-react-app": "^0.6.2",
-    "eslint-loader": "1.6.0",
-    "eslint-plugin-flowtype": "2.21.0",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-plugin-jsx-a11y": "4.0.0",
-    "eslint-plugin-react": "6.4.1",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.10.0",
-    "fs-extra": "0.30.0",
-    "html-webpack-plugin": "2.24.0",
-    "http-proxy-middleware": "0.17.3",
-    "jest": "18.1.0",
-    "json-loader": "0.5.4",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
+    "chalk": "2.4.1",
+    "connect-history-api-fallback": "1.5.0",
+    "cross-spawn": "6.0.5",
+    "css-loader": "0.28.11",
+    "detect-port": "1.2.3",
+    "dotenv": "6.0.0",
+    "eslint": "4.19.1",
+    "eslint-config-react-app": "^2.1.0",
+    "eslint-loader": "2.0.0",
+    "eslint-plugin-flowtype": "2.49.3",
+    "eslint-plugin-import": "2.12.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-react": "7.9.1",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.11",
+    "html-webpack-plugin": "3.2.0",
+    "http-proxy-middleware": "0.18.0",
+    "jest": "23.1.0",
+    "json-loader": "0.5.7",
     "object-assign": "4.1.1",
-    "postcss-loader": "1.2.2",
-    "promise": "7.1.1",
-    "react-dev-utils": "^0.5.2",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.14.0",
-    "webpack-dev-server": "1.16.2",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "2.0.2"
+    "postcss-loader": "2.1.5",
+    "promise": "8.0.1",
+    "react-dev-utils": "^5.0.1",
+    "style-loader": "0.21.0",
+    "url-loader": "1.0.1",
+    "webpack": "4.12.0",
+    "webpack-dev-server": "3.1.4",
+    "webpack-manifest-plugin": "2.0.3",
+    "whatwg-fetch": "2.0.4"
   },
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "build:components": "webpack  --config config/webpack-components.config.js",
-    "test": "node scripts/test.js --env=jsdom",
-    "postinstall": "postinstall-build build --script build:components"
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testPathIgnorePatterns": [
-      "<rootDir>[/\\\\](build|docs|node_modules|scripts)[/\\\\]"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web"
-    }
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   },
   "babel": {
     "presets": [
